Moved glue.qt.glue_toolbar to glue.viewers.common.qt.toolbar
