+++ glue/plugins/dendro_viewer/qt_widget.py
@@ -3,7 +3,7 @@ from __future__ import absolute_import, division, print_function
 from glue.external.qt import QtGui
 from glue import core
 from glue.plugins.dendro_viewer.client import DendroClient
-from glue.qt.glue_toolbar import GlueToolbar
+from glue.viewers.common.qt.toolbar import GlueToolbar
 from glue.qt.mouse_mode import PickMode
 from glue.qt.qtutil import load_ui
 from glue.utils.qt.widget_properties import (ButtonProperty, CurrentComboProperty,
