+++ glue/viewers/common/qt/tests/test_toolbar.py
@@ -4,9 +4,10 @@ from __future__ import absolute_import, division, print_function
 
 from glue.viewers.common.qt.mpl_widget import MplWidget
 
-from ..glue_toolbar import GlueToolbar
-from ..mouse_mode import MouseMode
-from ..qtutil import get_icon
+from glue.qt.mouse_mode import MouseMode
+from glue.qt.qtutil import get_icon
+
+from ..toolbar import GlueToolbar
 
 
 class MouseModeTest(MouseMode):
