+++ glue/viewers/common/qt/toolbar.py
