+++ glue/viewers/scatter/qt/viewer_widget.py
@@ -6,7 +6,7 @@ from glue.external.qt.QtCore import Qt
 from glue.external.qt import QtGui
 from glue import core
 from glue.viewers.scatter.client import ScatterClient
-from glue.qt.glue_toolbar import GlueToolbar
+from glue.viewers.common.qt.toolbar import GlueToolbar
 from glue.qt.mouse_mode import (RectangleMode, CircleMode,
                                 PolyMode, HRangeMode, VRangeMode)
 from glue.qt.qtutil import load_ui
